.github/tests/acceptance/test_scheduling/requirements.txt
pytest
rpi-lgpio
pytest-timeout